+++ partiql-lang/src/main/kotlin/org/partiql/lang/planner/transforms/LogicalResolvedToDefaultPhysicalVisitorTransform.kt
@@ -5,12 +5,6 @@ import org.partiql.lang.domains.PartiqlLogicalResolved
 import org.partiql.lang.domains.PartiqlLogicalResolvedToPartiqlPhysicalVisitorTransform
 import org.partiql.lang.domains.PartiqlPhysical
 import org.partiql.lang.errors.ProblemHandler
-import org.partiql.lang.planner.DML_COMMAND_FIELD_ACTION
-import org.partiql.lang.planner.DML_COMMAND_FIELD_ROWS
-import org.partiql.lang.planner.DML_COMMAND_FIELD_TARGET_ALIAS_INDEX
-import org.partiql.lang.planner.DML_COMMAND_FIELD_TARGET_UNIQUE_ID
-import org.partiql.lang.planner.DML_CONFLICT_ACTION_CONDITION
-import org.partiql.lang.planner.DmlAction
 
 /**
  * Transforms an instance of [PartiqlLogicalResolved.Statement] to [PartiqlPhysical.Statement],
@@ -169,53 +163,6 @@ internal class LogicalResolvedToDefaultPhysicalVisitorTransform(
                 i = DEFAULT_IMPL,
                 source = thiz.transformBexpr(node.source),
                 bindings = node.bindings.map { transformLetBinding(it) },
-                metas = node.metas
-            )
-        }
-    }
-
-    override fun transformStatementDml(node: PartiqlLogicalResolved.Statement.Dml): PartiqlPhysical.Statement {
-        val action = when (node.operation) {
-            is PartiqlLogicalResolved.DmlOperation.DmlInsert -> DmlAction.INSERT
-            is PartiqlLogicalResolved.DmlOperation.DmlDelete -> DmlAction.DELETE
-            is PartiqlLogicalResolved.DmlOperation.DmlReplace -> DmlAction.REPLACE
-            is PartiqlLogicalResolved.DmlOperation.DmlUpdate ->
-                TODO("DmlUpdate physical transform is not supported yet")
-        }.name.toLowerCase()
-
-        val condition = when (val op = node.operation) {
-            is PartiqlLogicalResolved.DmlOperation.DmlInsert -> null
-            is PartiqlLogicalResolved.DmlOperation.DmlDelete -> null
-            is PartiqlLogicalResolved.DmlOperation.DmlReplace -> op.condition?.let { transformExpr(it) }
-            is PartiqlLogicalResolved.DmlOperation.DmlUpdate -> op.condition?.let { transformExpr(it) }
-        }
-
-        val alias = when (val op = node.operation) {
-            is PartiqlLogicalResolved.DmlOperation.DmlInsert -> PartiqlPhysical.Expr.Lit(
-                op.targetAlias.index.toIonElement().asAnyElement()
-            )
-            is PartiqlLogicalResolved.DmlOperation.DmlDelete -> null
-            is PartiqlLogicalResolved.DmlOperation.DmlReplace -> PartiqlPhysical.Expr.Lit(
-                op.targetAlias.index.toIonElement().asAnyElement()
-            )
-            is PartiqlLogicalResolved.DmlOperation.DmlUpdate -> PartiqlPhysical.Expr.Lit(
-                op.targetAlias.index.toIonElement().asAnyElement()
-            )
-        }
-
-        val fields = mutableListOf(
-            structField(DML_COMMAND_FIELD_ACTION, action),
-            structField(DML_COMMAND_FIELD_TARGET_UNIQUE_ID, PartiqlPhysical.Expr.Lit(node.uniqueId.toIonElement().asAnyElement())),
-            structField(DML_COMMAND_FIELD_ROWS, transformExpr(node.rows)),
-        )
-
-        // Adds optional fields
-        alias?.let { fields.add(structField(DML_COMMAND_FIELD_TARGET_ALIAS_INDEX, alias)) }
-        condition?.let { fields.add(structField(DML_CONFLICT_ACTION_CONDITION, condition)) }
-
-        return PartiqlPhysical.build {
-            dmlQuery(
-                expr = struct(parts = fields, metas = node.metas),
                 metas = node.metas
             )
         }